Extract risk-based position sizing helper
